+++ apps/ledger-live-mobile/src/const/navigation.js
@@ -173,6 +173,7 @@ export const ScreenName = {
   CosmosEditMemo: "CosmosEditMemo",
   CosmosDelegationStarted: "CosmosDelegationStarted",
   CosmosDelegationValidator: "CosmosDelegationValidator",
+  CosmosDelegationValidatorSelect: "CosmosDelegationValidatorSelect",
   CosmosDelegationAmount: "CosmosDelegationAmount",
   CosmosDelegationSelectDevice: "CosmosDelegationSelectDevice",
   CosmosDelegationConnectDevice: "CosmosDelegationConnectDevice",
@@ -351,6 +352,7 @@ export const ScreenName = {
 
   // solana
   SolanaEditMemo: "SolanaEditMemo",
+  SolanaEditAmount: "SolanaEditAmount",
 
   PostBuyDeviceScreen: "PostBuyDeviceScreen",
 
@@ -416,6 +418,9 @@ export const NavigatorName = {
   PolkadotRebondFlow: "PolkadotRebondFlow",
   PolkadotNominateFlow: "PolkadotNominateFlow",
   PolkadotSimpleOperationFlow: "PolkadotSimpleOperationFlow",
+
+  // Solana
+  SolanaDelegationFlow: "SolanaDelegationFlow",
 
   NotificationCenter: "NotificationCenter",
   Market: "Market",