+++ tools/github-bot/src/tools.ts
@@ -5,7 +5,7 @@ export const isValidBranchName = (branch: string): boolean =>
   ["release", "hotfix"].includes(branch);
 
 export const isValidUser = (user: string): boolean =>
-  !["ledgerlive", "live-github-bot[bot]"].includes(user);
+  !["ledgerlive", "live-github-bot[bot]", "github-actions[bot]"].includes(user);
 
 export const isValidBody = (body: string | null): boolean => {
   if (!body) return false;
@@ -20,7 +20,7 @@ export const isValidBody = (body: string | null): boolean => {
     "### 🚀 Expectations to reach",
   ];
 
-  const results = body.split("\r\n").reduce(
+  const results = body.split(/\r?\n/).reduce(
     (acc, line) => {
       // Dummy description line has not been replaced.
       if (line === description) {
@@ -30,7 +30,9 @@ export const isValidBody = (body: string | null): boolean => {
         };
       }
 
-      const headingIndex = requiredHeadings.indexOf(line);
+      const headingIndex = requiredHeadings.findIndex((heading) =>
+        line.startsWith(heading)
+      );
       // Template required heading is still in the body.
       if (headingIndex > -1) {
         acc.matchHeadings[headingIndex] = true;